Added incrimental retry backoff strategy to handle servers that are using throttling. Also cleaned up debug logging.

+++ codeargos/webcrawler.py
@@ -1,5 +1,7 @@
 #!/usr/bin/env python3
 import requests
+from requests.adapters import HTTPAdapter
+from urllib3.util.retry import Retry
 import re
 import sys
 from urllib.parse import urlparse,urljoin
@@ -9,6 +11,7 @@ import signal
 from queue import Queue, Empty
 from concurrent.futures import ThreadPoolExecutor
 import logging
+import time
 
 class Scraper:
 
@@ -40,6 +43,7 @@ class Scraper:
             else:
                 return ""
         except Exception as ex:
+
             logging.exception(ex)
             return ""
         return parsed_html
@@ -69,6 +73,19 @@ class Scraper:
 
     def scrape(self):
         session = requests.session()
+
+        # Add incrimental backoff retry logic some we aren't slamming servers
+        retry_strategy = Retry(
+            total=10,
+            status_forcelist=[104,429,500,502,503,504],
+            method_whitelist=["HEAD", "GET", "OPTIONS"],
+            backoff_factor=1
+        )
+
+        adapter = HTTPAdapter(max_retries=retry_strategy)
+        session.mount('http://', adapter)
+        session.mount('https://', adapter)
+
         parsed_html = self.get_page(session, self.url)        
         if len(parsed_html) > 0:
             scraped_urls = self.get_links(self.url, parsed_html)
@@ -95,9 +112,12 @@ class WebCrawler:
         signal.signal(signal.SIGINT, self.dump_data)
 
     def dump_data(self, signal, frame):
-        with open('processed.txt', 'w') as f:
-            for item in self.processed_urls:
-                f.write("%s\n" % item)
+        choice = input( "\nEarly abort detected. Dump data already collected? (to processed.txt): [y/N] ")
+        choice = choice.lower()
+        if choice == 'y':
+            with open('processed.txt', 'w') as f:
+                for item in self.processed_urls:
+                    f.write("%s\n" % item)
         sys.exit()
 
     def process_scraper_results(self, future):
@@ -124,27 +144,27 @@ class WebCrawler:
         while True:
             try:
                 # get a url from the queue
-                target_url = self.queued_urls.get(timeout=60)
+                target_url = self.queued_urls.get(timeout=10)
 
                 # check that the url hasn't already been processed
                 if target_url not in self.processed_urls:
                     # add url to the processed list
                     self.processed_urls.add(target_url)
 
-                    logging.info(f'[URL] {target_url}')
+                    logging.debug(f'[URL] {target_url}')
 
                     job = self.pool.submit(Scraper(target_url).scrape)
                     job.add_done_callback(self.process_scraper_results)
 
                 if self.show_stats:
                     if i % LOG_EVERY_N == 0:
-                        print("Processed: {0:<8} | Queues: {1:<8} | Jobs: {2:<8}".format(
+                        print("Processed: {0:<8} | Queue: {1:<8} | Scheduled Jobs: {2:<8}".format(
                             len(self.processed_urls), 
                             self.queued_urls.qsize(),
                             self.pool._work_queue.qsize()))                    
                 i=i+1
             except Empty:
-                print("All done.")
+                logging.debug("All queues and jobs complete.")
                 break
             except Exception as e:
                 print(e)
